workers/dem/requirements.txt
numpy==1.26.4
rasterio==1.3.10
scipy==1.13.1
click==8.1.7
psycopg2-binary==2.9.9
python-dotenv==1.0.1
whitebox==2.3.5